requirements.txt
Django>=4.2.2, <4.3
djangorestframework>=3.14,<3.15
psycopg[binary]>=3.2, <3.3
drf-spectacular==0.28
python-dotenv==1.1.0
social-auth-app-django>=5.4.0,<=5.4.3
social-auth-core>=4.5.0,<=4.5.6
pillow>=11.0.0, <=11.2.1